library/requirements.txt
fastapi
uvicorn
nltk
pdfplumber
python-docx
